typed-protocols: tryRecvMessage
Added 'tryRecv' to 'Channel'.  It can be implemented either as
a non-blocking operation or, a read from a network interface with
a tight deadline.

Added 'tryRecvMessage' to 'Driver', which can terminate in the middle of
incremental decoding.  Both 'recvMessage' and 'tryRecvMessage' can
resume interrupted decoder.  If the underlying channel 'tryRecv' is
a non-blocking operation, then 'tryRecvMessage' will be non-blocking
too.

This patch also updates the implementation of 'runPeerWithDriver'.